Simplify service discovery with docker-compose
To simplify the demo repo we can use `links` in the docker-compose
file. This will allow each of the services to run alone and have compose
wire up the ports and addresses.
