+++ .gitignore
@@ -6,3 +6,4 @@ cmd/profile/profile
 cmd/rate/rate
 cmd/www/www
 cmd/search/search
+vendor/**
