+++ dialer/dialer.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 
 	"github.com/grpc-ecosystem/grpc-opentracing/go/otgrpc"
-	consul "github.com/hashicorp/consul/api"
-	lb "github.com/olivere/grpc/lb/consul"
 	opentracing "github.com/opentracing/opentracing-go"
 	"google.golang.org/grpc"
 )
@@ -17,17 +15,6 @@ type DialOption func(name string) (grpc.DialOption, error)
 func WithTracer(t opentracing.Tracer) DialOption {
 	return func(name string) (grpc.DialOption, error) {
 		return grpc.WithUnaryInterceptor(otgrpc.OpenTracingClientInterceptor(t)), nil
-	}
-}
-
-// WithBalancer enables client side load balancing
-func WithBalancer(cc *consul.Client) DialOption {
-	return func(name string) (grpc.DialOption, error) {
-		r, err := lb.NewResolver(cc, name, "")
-		if err != nil {
-			return nil, err
-		}
-		return grpc.WithBalancer(grpc.RoundRobin(r)), nil
 	}
 }
