+++ docker-compose.yml
@@ -1,15 +1,41 @@
 version: "3"
 services:
-  demo:
+  frontend:
     build: .
-    entrypoint: srv all
+    entrypoint: frontend
     ports:
       - "5000:5000"
     links:
       - jaeger
-      - consul
+      - search
+      - profile
     depends_on:
-      - consul
+      - search
+      - profile
+  search:
+    build: .
+    entrypoint: search
+    links:
+      - geo
+      - rate
+      - jaeger
+  profile:
+    build: .
+    entrypoint: profile
+    links:
+      - geo
+      - rate
+      - jaeger
+  geo:
+    build: .
+    entrypoint: geo
+    links:
+      - jaeger
+  rate:
+    build: .
+    entrypoint: rate
+    links:
+      - jaeger
   jaeger:
     image: jaegertracing/all-in-one:latest
     ports:
@@ -21,10 +47,3 @@ services:
       - "5775:5775/udp"
       - "6831:6831/udp"
       - "6832:6832/udp"
-  consul:
-    image: consul:latest
-    ports:
-      - "8300:8300"
-      - "8400:8400"
-      - "8500:8500"
-      - "8600:53/udp"
