+++ go.mod
@@ -1,0 +1,25 @@
+module github.com/harlow/go-micro-services
+
+go 1.12
+
+require (
+	github.com/apache/thrift v0.0.0-20161221203622-b2a4d4ae21c7 // indirect
+	github.com/codahale/hdrhistogram v0.0.0-20161010025455-3a0bb77429bd // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b // indirect
+	github.com/golang/protobuf v1.2.0
+	github.com/grpc-ecosystem/grpc-opentracing v0.0.0-20171214222146-0e7658f8ee99
+	github.com/hailocab/go-geoindex v0.0.0-20160127134810-64631bfe9711
+	github.com/opentracing-contrib/go-stdlib v0.0.0-20180308002341-f6b9967a3c69
+	github.com/opentracing/opentracing-go v1.0.2
+	github.com/prometheus/client_golang v0.9.2
+	github.com/stretchr/testify v1.3.0 // indirect
+	github.com/uber-go/atomic v1.4.0 // indirect
+	github.com/uber/jaeger-client-go v2.11.2+incompatible
+	github.com/uber/jaeger-lib v1.4.0 // indirect
+	go.uber.org/atomic v1.4.0 // indirect
+	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
+	golang.org/x/sync v0.0.0-20181221193216-37e7f081c4d4 // indirect
+	google.golang.org/genproto v0.0.0-20180306020942-df60624c1e9b // indirect
+	google.golang.org/grpc v1.10.0
+)
